+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - refactor http streaming server
   - rust 1.89
   - use ecow EcoString for the small strings that get cloned frequently
+  - replace strfmt with figura for dynamic string formatting
 
 - 1.12.11 (Aug 1 2025 dheijl)
   - more small refactorings